# api/docs/swagger.yaml
definitions:
  auth_service.AddClientRequest:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      login_strategy:
        type: integer
      login_table:
        $ref: '#/definitions/auth_service.LoginTable'
        type: object
      login_value:
        type: string
      password:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.ClientPrimaryKey:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
    type: object
...
  auth_service.CommonMessage:
    properties:
      data:
        type: string
      table_slug:
        type: string
    type: object
...
  auth_service.CreateClientPlatformRequest:
    properties:
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.GetClientPlatformListResponse:
    properties:
      client_platforms:
        items:
          $ref: '#/definitions/auth_service.ClientPlatform'
        type: array
      count:
        type: integer
    type: object
...
  auth_service.UpdateClientRequest:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      login_strategy:
        type: integer
      login_table:
        $ref: '#/definitions/auth_service.LoginTable'
        type: object
      login_value:
        type: string
      password:
        type: string
      project_id:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
  /client:
    delete:
      consumes:
      - application/json
      description: Get Client
      operationId: remove_client
      parameters:
      - description: RemoveClientBody
        in: body
        name: remove-client
        required: true
        schema:
          $ref: '#/definitions/auth_service.ClientPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Client
      tags:
      - Client
    post:
      consumes:
      - application/json
      description: Create Client
      operationId: create_client
      parameters:
      - description: AddClientRequestBody
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: Update Client
      operationId: update_client
      parameters:
      - description: UpdateClientRequestBody
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Client
      tags:
      - Client
  /client-platform:
    get:
      consumes:
      - application/json
      description: Get ClientPlatform List
      operationId: get_client_platform_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientPlatformListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetClientPlatformListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientPlatform List
      tags:
      - ClientPlatform
    post:
      consumes:
      - application/json
      description: Create ClientPlatform
      operationId: create_client_platform
      parameters:
      - description: CreateClientPlatformRequestBody
        in: body
        name: client-platform
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateClientPlatformRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ClientPlatform data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ClientPlatform
      tags:
      - ClientPlatform
    put:
      consumes:
      - application/json
      description: Update ClientPlatform
      operationId: update_client_platform
      parameters:
      - description: UpdateClientPlatformRequestBody
        in: body
        name: client-platform
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientPlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ClientPlatform data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ClientPlatform
      tags:
      - ClientPlatform
  /client-platform-detailed/{client-platform-id}:
    get:
      consumes:
      - application/json
      description: Get ClientPlatform By ID Detailed
      operationId: get_client_platform_detailed_by_id
      parameters:
      - description: client-platform-id
        in: path
        name: client-platform-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ClientPlatformDetailedBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientPlatform By ID Detailed
      tags:
      - ClientPlatform
  /client-platform/{client-platform-id}:
    delete:
      consumes:
      - application/json
      description: Get ClientPlatform
      operationId: delete_client_platform
      parameters:
      - description: client-platform-id
        in: path
        name: client-platform-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete ClientPlatform
      tags:
      - ClientPlatform
    get:
      consumes:
      - application/json
      description: Get ClientPlatform By ID
      operationId: get_client_platform_by_id
      parameters:
      - description: client-platform-id
        in: path
        name: client-platform-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ClientPlatformBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientPlatform By ID
      tags:
      - ClientPlatform
  /client-type:
    get:
      consumes:
      - application/json
      description: Get ClientType List
      operationId: get_client_type_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientTypeListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetClientTypeListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientType List
      tags:
      - ClientType
    post:
      consumes:
      - application/json
      description: Create ClientType
      operationId: create_client_type
      parameters:
      - description: CreateClientTypeRequestBody
        in: body
        name: client-type
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateClientTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ClientType data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientType'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ClientType
      tags:
      - ClientType
    put:
      consumes:
      - application/json
      description: Update ClientType
      operationId: update_client_type
      parameters:
      - description: UpdateClientTypeRequestBody
        in: body
        name: client-type
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ClientType data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientType'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ClientType
      tags:
      - ClientType
...
  /relation/{relation-id}:
    delete:
      consumes:
      - application/json
      description: Get Relation
      operationId: delete_relation
      parameters:
      - description: relation-id
        in: path
        name: relation-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Relation
      tags:
      - Relation
...
swagger: "2.0"